services/ml/requirements.txt
torch-geometric==2.5.0
torch-scatter==2.1.2
torch-sparse==0.6.18
torch-cluster==1.6.3
pytorch-lightning==2.2.0
wandb==0.16.3
mlflow==2.11.0
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.4.0
pydantic==2.5.3
protobuf==4.25.2
